feat: move tree parser into separate module
